requirements.txt
django
djangorestframework
gunicorn
django-heroku
psycopg2-binary==2.9.1
